move sequence expression into a new file
